Move toml constructor helper function to separate file
